Use libkdumpfile to read VMCOREINFO
If VMCOREINFO is found in the dump file itself, then libkdumpfile
has already parsed and stored it in attributes at initialization.
Let's use these attributes in this case.

However, makedumpfile has traditionally also allowed reading
VMCOREINFO data from external files. This is not directly supported
by libkdumpfile. The VMCOREINFO file must be read into memory and
used as a new value of the vmcoreinfo attribute (which gets reparsed
automatically). The original value of the attribute must be saved,
because it is used later to override values from the file.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>