eppic: Fix incompatible pointer type warnings
On Linux, eppic defines ull (a bit confusingly) as uint64_t. This
is an unsigned long on 64-bit platforms with both GCC and LLVM, and
the compiler complains about passing a pointer to an unsigned long
where a pointer to an unsigned long long is expected.

It is not a real bug, because both types are identical on 64-bit
platforms, and on 32-bit platforms, uint64_t denotes an unsigned
long long, but the warning is annoying.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>