Use libkdumpfile to read VMCOREINFO
If VMCOREINFO is found in the dump file itself, then libkdumpfile has
already parsed and stored it in attributes at initialization. Use these
attributes in this case.

However, makedumpfile has traditionally also allowed reading VMCOREINFO
data from an external file. This is not directly supported by libkdumpfile.
The external VMCOREINFO file content must be read into memory and used as a
new value of the vmcoreinfo attribute (which gets reparsed automatically).
The original value of the attribute must be also saved, because it is used
later to override values from the file.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>