Fix the --xen_phys_start option
Since Xen address translation now uses libkdumpfile, the value from
info->xen_phys_start is not used.

The default value of this attribute is initialized from Xen crash info by
libkdumpfile, so there is no need to repeat that logic in makedumpfile, but
if an explicit value was given on the command line, the corresponding
libkdump attribute must be set to that value.

Signed-off-by: Petr Tesarik <petr@tesarici.cz>